refactor sound code to use ECS
this is pretty terrible, oh well, let's push to master.

trying to integrate sounds with the entity component system. it's
now far more complicated and sketchy.

this is really messy because i don't have a resource manager and
i don't have a heap. so there's some terrible comptime stuff going
on. i think i'll add a resource manager and heap eventually so
that sounds become much simpler (just a single component type,
and synthesized sounds could just be referred to by a runtime name,
as if they were wav files).