split up "game" and "menu" input modes
this fixes the bug where, if you were at the in-game menu and hit the
spacebar to select "continue game", the player would fire a shot as well.

i created separate enums for game commands and menu commands, and added
an "input router" to pass along only the game command when no menu is
open, or only the menu command when the menu is open.

then i made some key binding changes: the "select menu option" is now
enter, not space. and "kill all monsters" is now backspace, not enter.