web: clear menu sounds whenever audio is disabled
this was a regression introduced when i removed menu sounds from the
component system. i had code to reset sounds when audio is disabled - but
only sounds in the component system. so this commit implements the same
thing for menu sounds.

(this is important because we don't want sounds triggered while audio is
disabled to all play at once when audio is turned on)