fix multithreading problems with menu sound effects
this is a prototype for how i'll deal with the same problem in game
sound effects, since the menu sound implementation is simpler.

the change here is that voice objects are split into main-thread fields
and audio-thread fields, with no overlap. there is a distinct "sync"
stage where information is copied from the one side to the other.