don't namespace math function inside structs
i think this style of programming creates cognitive load (when not using
intellisense, which i'm not...)

the dot operator is overloaded in many languages, where it could stand
for a constant offset, or a pointer indirection, or even a virtual
method lookup. also, the more dots in an expression, the harder it is to
read. for these reason i think `math.vec2Add` is superior to
`math.Vec2.add`.

and for some functions it's hard to decide which "class" it should be a
"method" of.

i also don't like method syntax for things like `a.add(b)` (where a new
object is returned, instead of `a` being mutated). this change makes
that not even a choice.

i can't get away from struct methods 100% in zig, but i can try to keep
it to a minimum.

i'm probably wrong, but no one will read this so it's ok.