apply removals immediately when resetting game
i believe the currently behavior of GBE is that entities that are
marked for removal are still alive (and visited by iterators)
until applyRemovals has been called.

so when we remove all ents as a part of resetting the game state
(ending a game), make sure to apply removals immediately.

this should fix the bug where a 1-up item was spawning after
starting a new game (the old PlayerController was still around,
with its score).

Closes #57