new mechanic: oxygen level ("tank")
you now have an oxygen level, which depletes over time. a bit restored
whenever you pick up a coin. it is reset to full when a new wave begins,
or when you die and respawn.

this should be something like the level timer in super mario bros.
there's more than enough time to get everything done. it's just there to
prevent the technical possibility of a game that never ends.