fix build.zig for zig 0.8.x
compile_zangscript is still commented out, i still need to update zangc
to compile with zig 0.8.

`zig build` now builds and installs everything. this seems to be the
default behavior of the zig build system. i don't know if it's good or
bad that i added all the different frontends as steps. i could have made
it a build argument (which frontend to use), and it only builds the one.
maybe that would make more sense?